requirements.txt
boto3>=1.34
pandas>=2.0
pyarrow>=14.0
yfinance>=0.2.36
requests>=2.31
pyyaml>=6.0
numpy>=1.24
feedparser>=6.0
beautifulsoup4>=4.12
lxml>=5.0
edgartools>=2.0
psycopg2-binary>=2.9
voyageai>=0.3
